chore: breakup fluxQueryBuiler.test
